Initial changes for GA-52/53

+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/data/UserRefreshQueueEntity.java
@@ -28,7 +28,7 @@ public class UserRefreshQueueEntity {
     private LocalDateTime lastUpdated;
 
     @Column(name = "access_types_min_version", nullable = false)
-    private Integer accessTypesMinVersion;
+    private Long accessTypesMinVersion;
 
     @Column(name = "deleted")
     private LocalDateTime deleted;
